filtering by genre id number
